requirements.txt
Django
djangorestframework
mysqlclient
python-decouple